Restructured SCMInvite related methods
