+++ scm/gitlab.go
@@ -325,17 +325,10 @@ func (*GitlabSCM) RequestReviewers(ctx context.Context, opt *RequestReviewersOpt
 	}
 }
 
-// AcceptRepositoryInvite implements the SCM interface
-func (*GitlabSCM) AcceptRepositoryInvites(_ context.Context, _ []*RepositoryInvitation) error {
+// AcceptRepositoryInvite implements the SCMInvite interface
+func (*GitlabSCM) AcceptRepositoryInvites(_ context.Context, _ *RepositoryInvitationOptions) error {
 	return ErrNotSupported{
 		SCM:    "gitlab",
-		Method: "AcceptRepositoryInvite",
-	}
-}
-
-func (*GitlabSCM) GetRepositoryInvites(_ context.Context, _ *RepositoryInvitationOptions) ([]*RepositoryInvitation, error) {
-	return nil, ErrNotSupported{
-		SCM:    "gitlab",
-		Method: "GetRepositoryInvites",
+		Method: "AcceptRepositoryInvites",
 	}
 }
